feat: impl `IntoIterator` and `Index` for `PmanFile`.
refactor: `parse_files`.

I swapped the mut state style to a iter.fold one. With this refactor, now I know
that the `Fmom` at the end of the file is actually bytes from the last file. It
is really odd that I didn't realize that later, but it was really a coincidace
of having those four characters at the end.