refactor(pman)!: copyright max size is now `55` bytes.
Decided to change the semantics of the file format spec, since forcing the user
to use a NULL terminated string is way much easier than trying to guess were the
copyright boundary ends; it also doesn't really affect the original pman file,
but I want to have a clean parser implementation? I will need to make a PR to
upstream to inform on the README that that is the "unofficial", but "technically
right" intention.

Also, I added some extra documentation to make explicit that String::len != # of
chars, to avoid future confusions.

refactor: use byte string literals instead of byte arrays.

refactor: rename `parse_*` functions to `read_*`.

Might change this back to `parse_*`, but for now I think this is a better
naming.

tests: include the full function name of the fuction being tested.

Before I was dropping the `parse_*`, which made reading test names kinda weird.